# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Book API",
    "description": "This is Book API",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
    "/authors": {
      "get": {
        "description": "Get all authors",
        "parameters": [
          {
            "name": "fullname",
            "in": "query",
            "type": "string"
          },
          {
            "name": "country",
            "in": "query",
            "type": "string"
          },
          {
            "name": "gender",
            "in": "query",
            "type": "string"
          },
          {
            "name": "birthdate",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Bad Request" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "post": {
        "description": "Create a new author",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "fullname": { "type": "string", "example": "John Doe" }
              }
            }
          }
        ],
        "responses": {
          "201": { "description": "Created" },
          "400": { "description": "Bad Request" },
          "409": { "description": "Conflict" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/authors/{id}": {
      "get": {
        "description": "Get author by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" }
        ],
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Bad Request" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "put": {
        "description": "Update author by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "fullname": { "type": "string", "example": "Jane Doe" }
              }
            }
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Bad Request" },
          "404": { "description": "Not Found" },
          "409": { "description": "Conflict" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "delete": {
        "description": "Delete author by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" }
        ],
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Bad Request" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/books": {
      "get": {
        "description": "Get all books",
        "parameters": [
          { "name": "title", "in": "query", "type": "string" },
          { "name": "authorId", "in": "query", "type": "string" }
        ],
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Bad Request" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "post": {
        "description": "Create a new book",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": { "type": "string", "example": "My First Book" },
                "authorId": { "type": "string", "example": "123" }
              }
            }
          }
        ],
        "responses": {
          "201": { "description": "Created" },
          "400": { "description": "Bad Request" },
          "409": { "description": "Conflict" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/books/{id}": {
      "get": {
        "description": "Get book by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" }
        ],
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Bad Request" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "put": {
        "description": "Update book by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": { "type": "string", "example": "Updated Book Title" },
                "authorId": { "type": "string", "example": "123" }
              }
            }
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Bad Request" },
          "404": { "description": "Not Found" },
          "409": { "description": "Conflict" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "delete": {
        "description": "Delete book by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" }
        ],
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Bad Request" },
          "404": { "description": "Not Found" },
          "500": { "description": "Internal Server Error" }
        }
      }
    }
  }
}